+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - Add source filename:line on repl exceptions (#812)
 - Add `defexception` special form (#819)
 - Add completion to REPL (#821)
+- Improve vector performance (#823)
 - Use PHPStan level 2
 
 ## [0.16.1](https://github.com/phel-lang/phel-lang/compare/v0.16.0...v0.16.1) - 2024-12-13